talkmetrix-backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
openai
python-multipart
